fix linux linting errors
