+++ cmd/finch/virtual_machine_settings_darwin_test.go
