+++ pkg/command/nerdctl.go
@@ -5,10 +5,6 @@ package command
 
 import (
 	"bytes"
-	"fmt"
-	"strings"
-
-	"golang.org/x/exp/slices"
 
 	"github.com/runfinch/finch/pkg/flog"
 	"github.com/runfinch/finch/pkg/system"
@@ -113,15 +109,4 @@ func (ncc *nerdctlCmdCreator) replaceBytes(s []byte, rs []Replacement) []byte {
 		s = bytes.ReplaceAll(s, []byte(r.Source), []byte(r.Target))
 	}
 	return s
-}
-
-func replaceOrAppend(orig []string, varName, newVar string) []string {
-	envIdx := slices.IndexFunc(orig, func(envVar string) bool {
-		return strings.HasPrefix(envVar, fmt.Sprintf("%s=", varName))
-	})
-
-	if envIdx != -1 {
-		return slices.Replace(orig, envIdx, envIdx+1, newVar)
-	}
-	return append(orig, newVar)
 }
