+++ pkg/config/lima_config_applier.go
@@ -1,9 +1,9 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
 package config
 
 import (
-	"github.com/spf13/afero"
-
-	"github.com/runfinch/finch/pkg/command"
 	"github.com/runfinch/finch/pkg/system"
 )
 
@@ -13,16 +13,6 @@ import (
 type LimaConfigApplierSystemDeps interface {
 	system.RuntimeArchGetter
 	system.RuntimeOSGetter
-}
-
-type limaConfigApplier struct {
-	cfg                    *Finch
-	cmdCreator             command.Creator
-	fs                     afero.Fs
-	limaDefaultConfigPath  string
-	limaOverrideConfigPath string
-	systemDeps             LimaConfigApplierSystemDeps
-	finchConfigPath        string
 }
 
 var _ LimaConfigApplier = (*limaConfigApplier)(nil)
